SIMPLE PATHFINDING!!!
note: when you create levels, try to keep them empty-ish (not to maze-y) because with the way pathfinding works enemies get stuck at walls trying to follow the player. that being said, if a level is maze-y, spawn the enemies ('e' in the text file) in room-ish places. also, spawn the player at a place where the walls will block enemy attacks to prevent getting killed on spawn.